refactor(datalab): ♻️ move _collect_results_from_issue_manager method to DataIssues class
Method extracted to class that is responsible for updating issues/summary/info
